backend/requirements.txt
fastapi[standard]
sqlalchemy
python-dotenv
uvicorn
psycopg2